\name{maxLik-internal}
\alias{print.summary.maxim}
\alias{print.summary.maxLik}
\alias{returnCode.maxim}
\alias{returnCode.maxLik}
% Document the following:

%%%%


\title{ Internal maxLik Functions }

\description{
   Internal maxLik Functions
}

\details{
   These are either various methods, or functions, not to be called
   directly by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
